Tweak packetizer -- make 32 slots
For the 16 X-pipeline setup, each pipeline
accepts 152 channels. Except this leads to packets which are >9kB.

Tweak the packetizer so that there are 32 slots of (up to) 128 channels.
With 2 slots to each X pipeline, this leads to 5kB packets.

This is easy to change at compile time for configurations with
other numbers of X-pipelines. But currently, packet size is
not runtime configurable.